requirements.txt
django==2.2.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0